move getUserById to UserService
