+++ src/main/java/com/joejoe2/chat/service/message/PublicMessageServiceImpl.java
@@ -10,8 +10,8 @@ import com.joejoe2.chat.models.PublicMessage;
 import com.joejoe2.chat.models.User;
 import com.joejoe2.chat.repository.channel.PublicChannelRepository;
 import com.joejoe2.chat.repository.message.PublicMessageRepository;
-import com.joejoe2.chat.repository.user.UserRepository;
 import com.joejoe2.chat.service.nats.NatsService;
+import com.joejoe2.chat.service.user.UserService;
 import com.joejoe2.chat.utils.ChannelSubject;
 import com.joejoe2.chat.validation.validator.MessageValidator;
 import com.joejoe2.chat.validation.validator.PageRequestValidator;
@@ -27,7 +27,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class PublicMessageServiceImpl implements PublicMessageService {
-  @Autowired UserRepository userRepository;
+  @Autowired UserService userService;
   @Autowired PublicChannelRepository channelRepository;
   @Autowired PublicMessageRepository messageRepository;
   @Autowired NatsService natsService;
@@ -36,13 +36,6 @@ public class PublicMessageServiceImpl implements PublicMessageService {
   MessageValidator messageValidator = MessageValidator.getInstance();
 
   PageRequestValidator pageValidator = PageRequestValidator.getInstance();
-
-  private User getUserById(String userId) throws UserDoesNotExist {
-    return userRepository
-        .findById(uuidValidator.validate(userId))
-        .orElseThrow(
-            () -> new UserDoesNotExist("user with id=%s does not exist !".formatted(userId)));
-  }
 
   private PublicChannel getChannelById(String channelId) throws ChannelDoesNotExist {
     return channelRepository
@@ -57,7 +50,7 @@ public class PublicMessageServiceImpl implements PublicMessageService {
   @Transactional(rollbackFor = Exception.class)
   public PublicMessageDto createMessage(String fromUserId, String channelId, String message)
       throws UserDoesNotExist, ChannelDoesNotExist {
-    User user = getUserById(fromUserId);
+    User user = userService.getUserById(fromUserId);
     PublicChannel channel = getChannelById(channelId);
 
     PublicMessage publicMessage =
