+++ src/main/java/com/joejoe2/chat/service/user/UserService.java
@@ -1,18 +1,30 @@
-package com.joejoe2.chat.service.user.auth;
+package com.joejoe2.chat.service.user;
 
 import com.joejoe2.chat.data.UserDetail;
+import com.joejoe2.chat.exception.UserDoesNotExist;
 import com.joejoe2.chat.models.User;
 import com.joejoe2.chat.repository.user.UserRepository;
+import com.joejoe2.chat.validation.validator.UUIDValidator;
 import java.util.UUID;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
-public class UserDetailService implements UserDetailsService {
+public class UserService implements UserDetailsService {
   @Autowired UserRepository userRepository;
+  UUIDValidator uuidValidator = UUIDValidator.getInstance();
+
+  @Transactional
+  public User getUserById(String userId) throws UserDoesNotExist {
+    return userRepository
+        .findById(uuidValidator.validate(userId))
+        .orElseThrow(
+            () -> new UserDoesNotExist("user with id=%s does not exist !".formatted(userId)));
+  }
 
   @Override
   public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
